refactor(account-storage): move ServerSettingsSerializer to feature:account:storage:legacy
