+++ feature/account/storage/legacy/build.gradle.kts
@@ -1,0 +1,19 @@
+plugins {
+    id(ThunderbirdPlugins.Library.android)
+}
+
+android {
+    namespace = "net.thunderbird.feature.account.storage.legacy"
+}
+
+dependencies {
+    implementation(projects.core.preferences)
+
+    implementation(projects.mail.common)
+
+    implementation(projects.legacy.account)
+
+    implementation(libs.moshi)
+
+    testImplementation(projects.mail.protocols.imap)
+}
