+++ legacy/core/src/main/java/com/fsck/k9/AccountPreferenceSerializer.kt
@@ -25,8 +25,8 @@ import app.k9mail.legacy.account.ShowPictures
 import app.k9mail.legacy.account.SortType
 import com.fsck.k9.helper.Utility
 import com.fsck.k9.preferences.StorageEditor
-import net.thunderbird.core.mail.folder.api.SpecialFolderSelection
 import net.thunderbird.core.preferences.Storage
+import net.thunderbird.feature.account.storage.legacy.ServerSettingsSerializer
 import net.thunderbird.feature.notification.NotificationLight
 import net.thunderbird.feature.notification.NotificationSettings
 import net.thunderbird.feature.notification.NotificationVibration
