+++ legacy/core/src/main/java/com/fsck/k9/KoinModule.kt
@@ -11,6 +11,7 @@ import com.fsck.k9.mail.ssl.TrustedSocketFactory
 import com.fsck.k9.mailstore.LocalStoreProvider
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.GlobalScope
+import net.thunderbird.feature.account.storage.legacy.ServerSettingsSerializer
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
