+++ legacy/core/src/main/java/com/fsck/k9/preferences/ServerSettingsWriter.kt
@@ -1,6 +1,5 @@
 package com.fsck.k9.preferences
 
-import com.fsck.k9.ServerSettingsSerializer
 import com.fsck.k9.mail.AuthType
 import com.fsck.k9.mail.ConnectionSecurity
 import com.fsck.k9.mail.ServerSettings
@@ -11,6 +10,7 @@ import com.fsck.k9.preferences.ServerSettingsDescriptions.HOST
 import com.fsck.k9.preferences.ServerSettingsDescriptions.PASSWORD
 import com.fsck.k9.preferences.ServerSettingsDescriptions.PORT
 import com.fsck.k9.preferences.ServerSettingsDescriptions.USERNAME
+import net.thunderbird.feature.account.storage.legacy.ServerSettingsSerializer
 
 internal class ServerSettingsWriter(
     private val serverSettingsSerializer: ServerSettingsSerializer,
