+++ legacy/storage/src/main/java/com/fsck/k9/preferences/migration/StorageMigrationTo14.kt
@@ -2,7 +2,7 @@ package com.fsck.k9.preferences.migration
 
 import android.database.sqlite.SQLiteDatabase
 import app.k9mail.core.common.mail.Protocols
-import com.fsck.k9.ServerSettingsSerializer
+import net.thunderbird.feature.account.storage.legacy.ServerSettingsSerializer
 
 /**
  * Rewrite 'folderPushMode' value of non-IMAP accounts to 'NONE'.
